requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
python-frontmatter>=1.0.1
markdown>=3.5.0
